Fix WDTimer bug

+++ BSP/Simulator/DataGeneration/WDTimer.py
@@ -1,5 +1,6 @@
 import csv
 import os
+import os.path
 import fcntl
 import config
 
@@ -16,44 +17,46 @@ reset = '2'
 def Check_State():
     global duration
     global TimerEnable
-    if TimerEnable == True:
+    if os.path.isfile(file) == True:    #only execute the following code if the .csv file has been created
         os.makedirs(os.path.dirname(file), exist_ok=True)
-        with open(file, 'r+') as csvfile:
-            fcntl.flock(csvfile.fileno(), fcntl.LOCK_EX)  
-            csvreader = csv.reader(csvfile)
-            currentState = next(csvreader)
-            if ((currentState[0] == reset) and (duration > 0)):
-                duration = 15 #reset countdown
-                csvfile.seek(0)
-                csvwriter = csv.writer(csvfile)
-                csvwriter.writerow([" "]) #Kick watchdog. Clear the reset message ('2').
-                csvwriter.writerow(["0"])
-            elif (duration <= 0):
-                csvfile.seek(0)
-                csvwriter = csv.writer(csvfile)
-                csvwriter.writerow([" ", "b"]) #print system reset message. The ' ' has no significance, it's just there to keep the formatting. The 'b' notifies the C program that the system has been reset
-                csvwriter.writerow(["0"])
-                os.kill(int(pid), 9) #system reset - kill process
-                TimerEnable = False #After system reset, watchdog can start again 
-                duration = 15 
-            fcntl.flock(csvfile.fileno(), fcntl.LOCK_UN)
+        if TimerEnable == True:
+            with open(file, 'r+') as csvfile:
+                fcntl.flock(csvfile.fileno(), fcntl.LOCK_EX)  
+                csvreader = csv.reader(csvfile)
+                currentState = next(csvreader)
+                if ((currentState[0] == reset) and (duration > 0)):
+                    duration = 15 #reset countdown
+                    csvfile.seek(0)
+                    csvwriter = csv.writer(csvfile)
+                    csvwriter.writerow([" "]) #Kick watchdog. Clear the reset message ('2').
+                    csvwriter.writerow(["0"])
+                elif (duration <= 0):
+                    csvfile.seek(0)
+                    csvwriter = csv.writer(csvfile)
+                    csvwriter.writerow([" ", "b"]) #print system reset message. The ' ' has no significance, it's just there to keep the formatting. The 'b' notifies the C program that the system has been reset
+                    csvwriter.writerow(["0"])
+                    os.kill(int(pid), 9) #system reset - kill process
+                    TimerEnable = False #After system reset, watchdog can start again 
+                    duration = 15 
+                fcntl.flock(csvfile.fileno(), fcntl.LOCK_UN)
 
 
 def WD_Enable():
     global TimerEnable
     global pid
-    os.makedirs(os.path.dirname(file), exist_ok=True)
-    with open(file, 'r') as csvfile:
-        fcntl.flock(csvfile.fileno(), fcntl.LOCK_EX)  
-        csvreader = csv.reader(csvfile)
-        currentState = next(csvreader)
-        if TimerEnable == False:    #check for timer startup
-            if currentState[0] == start:
-                    TimerEnable = True #start Watchdog timer
-            secondLine = next(csvreader)
-            if secondLine[0] != 0:
-                pid = secondLine[0]
-        fcntl.flock(csvfile.fileno(), fcntl.LOCK_UN)
+    if os.path.isfile(file) == True:    #only execute the following code if the .csv file has been created
+        os.makedirs(os.path.dirname(file), exist_ok=True)
+        with open(file, 'r') as csvfile:
+            fcntl.flock(csvfile.fileno(), fcntl.LOCK_EX)  
+            csvreader = csv.reader(csvfile)
+            currentState = next(csvreader)
+            if TimerEnable == False:    #check for timer startup
+                if currentState[0] == start:
+                        TimerEnable = True #start Watchdog timer
+                secondLine = next(csvreader)
+                if secondLine[0] != 0:
+                    pid = secondLine[0]
+            fcntl.flock(csvfile.fileno(), fcntl.LOCK_UN)
 
 def Tick():
     global duration
